+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- Add ability to connect to a Wi-Fi network
+- Updated the executable to connect to a Wi-Fi network
+
 ## 0.0.1
 
 - Initial version.